fix(sourcemaps): Fix double association bug (#2764)
Fixes a bug where the same sourcemap can be associated with mulitple
source files.

### Change description

We solve this problem by keeping track of the sourcemaps we have already
associated with a source file. If we guess a sourcemap which has already
been associated, we do not associate the given source file with any
sourcemap.

Also, we iterate first through all the sources which explicitly
reference the sourcemap – i.e., for which, we don't have to guess the
sourcemap. These are associated first. As they are explicitly linked to
a sourcemap, we intentionally **do not prevent duplicates** here.

On the second iteration, we guess the sourcemap reference for all
remaining sources. **We enforce the no duplicates rule on the second
iteration**. Sources cannot be associated with a sourcemap that was
explicitly referenced by another source on the first iteration. If
multiple sources guess the same sourcemap, and that sourcemap is not
explicitly associated with any source, then none of the sources will be
associated with that sourcemap (as it is unclear how that sourcemap
should be associated).

In all cases, we warn the users that we failed to associate the sources
properly.

### Issues

Fixes #2445
Fixes
[CLI-5](https://linear.app/getsentry/issue/CLI-5/ensure-collect-sourcemap-references-does-not-associate-the-same)