feat(proguard): Remove release association arguments (#2876)
⚠️ **Breaking change:** Do not merge until ready to release in a major.

Remove the `--app-id`, `--version`, and `--version-code` arguments to the `sentry-cli upload-proguard` command. Users currently using these arguments should simply stop using them. With this change, we also no longer wait for assembly of the mapping to finish on the backend; previously we needed to wait to associate the release.

This change also removes any code that had been called when these flags were specified.

 - Resolves #2868
 - Resolves [CLI-197](https://linear.app/getsentry/issue/CLI-197/remove-flags-for-setting-release)