chore(preprod): Add AssetCatalog Parsing support for Multisize sets (#2879)
Certain assets weren't getting picked up for the Firefox app, and the
ones missing appeared to be multisize sets, like this

```
    {
        "vector": false,
        "type": 0,
        "imageId": "42D31467-E63B-49D8-8D9B-1B608059E344",
        "size": 236,
        "height": 1024,
        "name": "AppIcon_Alt_Color_Purple",
        "filename": "AppIcon_Alt_Color_Purple",
        "idiom": "phone",
        "width": 1024
    },
```

They won't have a a file extension either it seems (ex.
`Assets.car/AppIcon_Alt_Color_Purple`). So we'll go ahead and save them
as a (default type) png