ref(uploads): Split `build_artifact_bundle` into separate module
Move `build_artifact_bundle` (and functions only it calls) out of `crate::utils::file_upload` and into a new module called `crate::utils::source_bundle`, also renaming the function to `build`.

We are also moving terminology to "source bundle," as these are **not** the artifact bundles for JavaScript and should not be confused with them.

This change only moves code around and renames items; there are no other code changes. Those come in subsequent PRs.