ref(bundles): Introduce `BundleContext` struct (#2923)
### Description
As we don't need all of the fields of the `UploadContext` struct for `source_bundle::build` (in particular, we don't need the `chunk_upload_options`), we introduce a new struct with only the fields needed for the function.

We will use this new struct in the next PR to avoid needing to get chunk upload options for the `bundle-jvm` command.

### Issues
- See #2927
- See [CLI-220](https://linear.app/getsentry/issue/CLI-220/dont-call-chunk-upload-for-bundle-jvm)