fix(sourcemaps): Fix incorrect chunk upload not supported error (#2931)
Supplying a non-existent organization to the `sentry-cli sourcemaps upload` can cause an error, which states "This version of Sentry does not support artifact bundles ...", even though this is not true. The reason is that the endpoint to get the chunk upload options contains the organization, and if a non-existent organization is supplied, we get a 404 error.

Here, we fix the error message to indicate that an incorrect organization is the most likely explanation, while acknowledging that lacking support for artifact bundles could also cause the problem.

- Resolves #2820
- Resolves [CLI-176](https://linear.app/getsentry/issue/CLI-176/non-existent-org-causes-chunk-upload-not-supported-error)