chore: Prepare to use TypeScript natively (#2910)
### Description

This PR is a starting point of upgrading to TypeScript and preparing for
future upgrades around testing and maintenance.

I've seen that couple of types between the `*.js` files and the
`index.d.ts` were not 100% in sync, which is one of the reasons the move
to TypeScript natively could help here.

### What's new / changed

- moved `js` to `lib` in order to keep `js` as the same output folder
which is delivered to NPM
- `js/index.d.ts` stays as type export for NPM and the types are
re-exported via the jsdoc `@typedef`
- `types.ts` is actually the same as the previous `index.d.ts` excluding
the `SentryCli` class
- TypeScript is partially used to keep the changes minimal
- I aligned the jsdoc types with the actual text from the previous
`index.d.ts`