ref: Use undici to download binary (#2993)
This removes `node-fetch` entirely. We could have upgrade to `node-fetch@3`, but this one requires ESM-only, which we can't support (we still need CJS support for our `@sentry/*` bundler packages).

With this PR we use now `undici` which also has a `ProxyAgent` included, so we got rid of `https-proxy-agent` as well.

Closes #1810
Closes [CLI-25](https://linear.app/getsentry/issue/CLI-25/upgrade-to-node-fetch-v3-to-avoid-punycode-deprecation-warning)