chore: Remove install script in GitHub Actions (#3020)
This script is actually not needed for our testing purpose. This was
initially in #2910 added, since we call `--ignore-scripts` before, which
doesn't call the `postinstall` script (so it was mainly added to keep
the same functionality as before).

This line actually also caused issues for our release process - that
should be fixed now